UPDATE Firebase Authentication
Cambios Realizados:
    1) Se implementó totalmente el flujo de autentificación de usuarios en la aplicación.

+++ app/src/main/java/org/equiposeis/huellitasaventureras/dataModels/Persona.kt
@@ -4,7 +4,7 @@ open class Persona(
     var nombre:String,
     var genero:Int,
     var edad:Int,
-    var numero_telefonico:Int,
+    var numero_telefonico:Long,
     var domicilio:String,
     var correo_electronico:String
 )
